@@ -4,60 +4,52 @@
   "author": "yosiakatuki",
   "license": "GPLv2",
   "devDependencies": {
-	"@babel/core": "^7.12.10",
-	"@babel/preset-env": "^7.12.11",
-	"@wordpress/browserslist-config": "^2.7.0",
-	"@wordpress/eslint-plugin": "^9.0.4",
-	"@wordpress/prettier-config": "^1.0.3",
-	"autoprefixer": "^10.2.5",
-	"babel-loader": "^8.2.2",
-	"babel-preset-minify": "^0.5.1",
-	"browserslist": "^4.16.0",
-	"css-declaration-sorter": "^6.0.3",
-	"css-mqpacker": "^6.0.2",
-	"css-vars-ponyfill": "^2.4.2",
-	"cssnano": "^5.0.2",
-	"del": "^5.1.0",
-	"eslint": "^7.26.0",
-	"eslint-config-prettier": "^7.2.0",
-	"eslint-plugin-prettier": "^3.4.0",
-	"gulp": "^4.0.2",
-	"gulp-babel": "^8.0.0",
-	"gulp-postcss": "^9.0.0",
-	"gulp-rename": "^2.0.0",
-	"gulp-sass": "^4.1.0",
-	"gulp-zip": "^4.2.0",
-	"node-sass": "^4.14.1",
-	"node-sass-package-importer": "^5.2.0",
-	"normalize.css": "^8.0.1",
-	"object-fit-images": "^3.2.4",
-	"postcss": "^8.2.15",
-	"postcss-clean": "^1.1.0",
-	"postcss-cli": "^8.3.1",
-	"prettier": "npm:wp-prettier@^2.2.1-beta-1",
-	"rimraf": "^2.7.1",
-	"simple-icons": "^2.19.0",
-	"stickyfilljs": "^2.1.0",
-	"vue": "^2.6.12",
-	"webpack": "^4.44.2",
-	"webpack-stream": "^5.2.1"
+    "@babel/core": "^7.15.0",
+    "@babel/preset-env": "^7.15.0",
+    "@wordpress/browserslist-config": "^4.1.0",
+    "@wordpress/eslint-plugin": "^9.1.0",
+    "@wordpress/prettier-config": "^1.1.0",
+    "autoprefixer": "^10.3.1",
+    "babel-loader": "^8.2.2",
+    "browserslist": "^4.16.7",
+    "css-declaration-sorter": "^6.1.1",
+    "css-vars-ponyfill": "^2.4.6",
+    "cssnano": "^5.0.7",
+    "del": "^6.0.0",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^3.4.0",
+    "gulp": "^4.0.2",
+    "gulp-babel": "^8.0.0",
+    "gulp-postcss": "^9.0.0",
+    "gulp-rename": "^2.0.0",
+    "gulp-sass": "^5.0.0",
+    "gulp-zip": "^5.1.0",
+    "object-fit-images": "^3.2.4",
+    "postcss": "^8.3.6",
+    "postcss-clean": "^1.2.2",
+    "postcss-cli": "^8.3.1",
+    "prettier": "npm:wp-prettier@^2.2.1-beta-1",
+    "rimraf": "^3.0.2",
+    "sass": "^1.37.5",
+    "simple-icons": "^5.9.0",
+    "stickyfilljs": "^2.1.0",
+    "webpack": "^5.50.0",
+    "webpack-stream": "^6.1.2"
   },
   "scripts": {
-	"zip": "gulp zip",
-	"watch": "gulp",
-	"csscomb": "csscomb src/sass",
-	"lint-js": "eslint ./src/js/*.js",
-	"lint-js:fix": "eslint --fix ./src/js/*.js",
-	"build:js": "gulp js",
-	"build:sass": "gulp sass",
-	"build:deploy-files": "gulp createDeployFiles",
-	"build:deploy-files-beta": "gulp createDeployFilesBeta",
-	"build:icons": "node ./src/js/icons/brand.js",
-	"build:pot": "wp i18n make-pot . languages/ystandard.pot --domain=\"ystandard\" --exclude=\"library,tests,vendor,bin\"",
-	"build": "npm run build:icons && gulp build",
-	"build:theme-zip": "gulp createDeployFiles && gulp cleanTemp"
+    "watch": "gulp",
+    "lint-js": "eslint ./src/js/*.js",
+    "lint-js:fix": "eslint --fix ./src/js/*.js",
+    "build:js": "gulp js",
+    "build:sass": "gulp sass",
+    "build:icons": "node ./src/js/icons/brand.js",
+    "build": "npm run build:icons && gulp build",
+    "i18n:pot": "wp i18n make-pot . languages/ystandard.pot --domain=\"ystandard\" --exclude=\"library,tests,vendor,bin\"",
+    "production": "gulp production",
+    "production:beta": "gulp productionBeta"
   },
   "browserslist": [
-	"extends @wordpress/browserslist-config"
+    "extends @wordpress/browserslist-config"
   ]
 }
