@@ -1,1 +1,1 @@
-"use strict";function copyClipboardForm(a){var b=jQuery(a).prev(),c=jQuery(a).next();b.select(),document.execCommand("copy"),c.length&&c.show().fadeOut(2e3)}document.addEventListener("DOMContentLoaded",function(){var a=jQuery(".copy-form__button:not(.is-without-event)");a.length&&a.each(function(a,b){b.addEventListener("click",function(a){a.preventDefault(),copyClipboardForm(a.target)})})});
+"use strict";function copyClipboardForm(target){var text=jQuery(target).prev();var info=jQuery(target).next();text.select();document.execCommand("copy");if(info.length){info.show().fadeOut(2000)}}document.addEventListener("DOMContentLoaded",function(){var copyButtons=jQuery(".copy-form__button:not(.is-without-event)");if(copyButtons.length){copyButtons.each(function(i,elem){elem.addEventListener("click",function(e){e.preventDefault();copyClipboardForm(e.target)})})}});
