@@ -1,1 +1,1 @@
-"use strict";jQuery(document).ready(function(a){var b=null,c="is-selected",d=a(".ys-custom-uploader__hidden").val();a(".ys-custom-uploader__clear").click(function(){a(".ys-custom-uploader__hidden").val(""),a(".ys-custom-uploader__preview").text("\u753B\u50CF\u304C\u9078\u629E\u3055\u308C\u3066\u307E\u305B\u3093\u3002"),a(".ys-custom-uploader").toggleClass(c)}),a(".ys-custom-uploader__select").click(function(f){return f.preventDefault(),b?void b.open():void(b=wp.media({title:"\u753B\u50CF\u3092\u9078\u629E",button:{text:"\u9078\u629E"},multiple:!1}),b.on("select",function(){var e=b.state().get("selection");e.each(function(b){d=b.toJSON().url,a(".ys-custom-uploader__hidden").val(d),a(".ys-custom-uploader__preview").text("").append("<img src=\""+d+"\" />"),a(".ys-custom-uploader").toggleClass(c)})}),b.open())}),d&&(console.log(d),a(".ys-custom-uploader__preview").text("").append("<img src=\""+d+"\" />"),a(".ys-custom-uploader").addClass(c))});
+"use strict";jQuery(document).ready(function($){var custom_uploader=null;var classSelected="is-selected";var imageUrl=$(".ys-custom-uploader__hidden").val();$(".ys-custom-uploader__clear").click(function(e){$(".ys-custom-uploader__hidden").val("");$(".ys-custom-uploader__preview").text("\u753B\u50CF\u304C\u9078\u629E\u3055\u308C\u3066\u307E\u305B\u3093\u3002");$(".ys-custom-uploader").toggleClass(classSelected)});$(".ys-custom-uploader__select").click(function(e){e.preventDefault();if(custom_uploader){custom_uploader.open();return}custom_uploader=wp.media({title:"\u753B\u50CF\u3092\u9078\u629E",button:{text:"\u9078\u629E"},multiple:false});custom_uploader.on("select",function(){var image=custom_uploader.state().get("selection");image.each(function(file){imageUrl=file.toJSON().url;$(".ys-custom-uploader__hidden").val(imageUrl);$(".ys-custom-uploader__preview").text("").append("<img src=\""+imageUrl+"\" />");$(".ys-custom-uploader").toggleClass(classSelected)})});custom_uploader.open()});if(imageUrl){console.log(imageUrl);$(".ys-custom-uploader__preview").text("").append("<img src=\""+imageUrl+"\" />");$(".ys-custom-uploader").addClass(classSelected)}});
