@@ -1,1 +1,1 @@
-"use strict";(function(a){wp.customize.bind("ready",function(){var b=this;a("#_customize-input-background_repeat,#_customize-input-background_attachment,#_customize-input-background_size,[name=\"background-position\"]").change(function(){setBackgroundPresetCustom(a)})})})(jQuery);function setBackgroundPresetCustom(a){var b=a("#_customize-input-background_preset");"default"===b.val()&&b.val("custom")}
+"use strict";(function($){wp.customize.bind("ready",function(){var customize=this;var selector="#_customize-input-background_repeat,#_customize-input-background_attachment,#_customize-input-background_size,[name=\"background-position\"]";$(selector).change(function(){setBackgroundPresetCustom($)})})})(jQuery);function setBackgroundPresetCustom($){var preset=$("#_customize-input-background_preset");if("default"===preset.val()){preset.val("custom")}}
