@@ -1,1 +1,1 @@
-"use strict";var fixedHeader=function(a){if(a(".has-fixed-header").length){var b=a(".site-header");if(0===b.length)return;b.prepend("<span class=\"header-height-info\">\u30D8\u30C3\u30C0\u30FC\u9AD8\u3055:<span class=\"header-height-num\"></span>px</span>"),a(".header-height-num").text(b.outerHeight()),a(window).on("resize",function(){a(".header-height-num").text(" ... "),setTimeout(function(){a(".header-height-num").text(b.outerHeight())},800)})}};document.addEventListener("DOMContentLoaded",function(){fixedHeader(jQuery)});
+"use strict";var fixedHeader=function fixedHeader($){if($(".has-fixed-header").length){var header=$(".site-header");if(0===header.length){return}header.prepend("<span class=\"header-height-info\">\u30D8\u30C3\u30C0\u30FC\u9AD8\u3055:<span class=\"header-height-num\"></span>px</span>");$(".header-height-num").text(header.outerHeight());$(window).on("resize",function(){$(".header-height-num").text(" ... ");setTimeout(function(){$(".header-height-num").text(header.outerHeight())},800)})}};document.addEventListener("DOMContentLoaded",function(){fixedHeader(jQuery)});
