@@ -1,1 +1,1 @@
-"use strict";var _vue=_interopRequireDefault(require("vue"));function _interopRequireDefault(a){return a&&a.__esModule?a:{default:a}}new _vue.default({el:"#ys-search-icons",data:{keyword:"",icons:searchIcons},methods:{copy:function copy(a,b){this.$refs[a]&&(this.$refs[b][0].classList.remove("is-show"),this.$refs[a][0].focus(),this.$refs[a][0].select(),document.execCommand("copy"),this.$refs[b][0].classList.add("is-show"))}},computed:{filteredIcons:function filteredIcons(){var a=[];if(!this.keyword.trim())return this.icons;for(var b in this.icons){var c=this.icons[b];-1!==c.name.indexOf(this.keyword)&&a.push(c)}return a}}});
+"use strict";document.addEventListener("DOMContentLoaded",function(){var search=document.getElementById("icon-search");search.addEventListener("keyup",function(e){var $=jQuery;var inputValue=$(e.target).val();$(".ys-icon-search__item").each(function(index,element){if(!inputValue){$(element).css("display","block")}else{var iconName=$(element).data("icon-name");var display=-1===iconName.indexOf(inputValue)?"none":"block";$(element).css("display",display)}})})});
