@@ -1,173 +1,2 @@
-!function(e){var t={};function r(n){if(t[n])return t[n].exports;var o=t[n]={i:n,l:!1,exports:{}};return e[n].call(o.exports,o,o.exports,r),o.l=!0,o.exports}r.m=e,r.c=t,r.d=function(e,t,n){r.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:n})},r.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},r.t=function(e,t){if(1&t&&(e=r(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var n=Object.create(null);if(r.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var o in e)r.d(n,o,function(t){return e[t]}.bind(null,o));return n},r.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return r.d(t,"a",t),t},r.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},r.p="",r(r.s=7)}({1:function(e,t,r){"use strict";
-/*! npm.im/object-fit-images 3.2.4 */var n="bfred-it:object-fit-images",o=/(object-fit|object-position)\s*:\s*([-.\w\s%]+)/g,s="undefined"==typeof Image?{style:{"object-position":1}}:new Image,a="object-fit"in s.style,i="object-position"in s.style,c="background-size"in s.style,u="string"==typeof s.currentSrc,l=s.getAttribute,f=s.setAttribute,d=!1;function p(e,t,r){var n="data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='"+(t||1)+"' height='"+(r||0)+"'%3E%3C/svg%3E";l.call(e,"src")!==n&&f.call(e,"src",n)}function m(e,t){e.naturalWidth?t(e):setTimeout(m,100,e,t)}function v(e){var t=function(e){for(var t,r=getComputedStyle(e).fontFamily,n={};null!==(t=o.exec(r));)n[t[1]]=t[2];return n}(e),r=e[n];if(t["object-fit"]=t["object-fit"]||"fill",!r.img){if("fill"===t["object-fit"])return;if(!r.skipTest&&a&&!t["object-position"])return}if(!r.img){r.img=new Image(e.width,e.height),r.img.srcset=l.call(e,"data-ofi-srcset")||e.srcset,r.img.src=l.call(e,"data-ofi-src")||e.src,f.call(e,"data-ofi-src",e.src),e.srcset&&f.call(e,"data-ofi-srcset",e.srcset),p(e,e.naturalWidth||e.width,e.naturalHeight||e.height),e.srcset&&(e.srcset="");try{!function(e){var t={get:function(t){return e[n].img[t||"src"]},set:function(t,r){return e[n].img[r||"src"]=t,f.call(e,"data-ofi-"+r,t),v(e),t}};Object.defineProperty(e,"src",t),Object.defineProperty(e,"currentSrc",{get:function(){return t.get("currentSrc")}}),Object.defineProperty(e,"srcset",{get:function(){return t.get("srcset")},set:function(e){return t.set(e,"srcset")}})}(e)}catch(e){window.console&&console.warn("https://bit.ly/ofi-old-browser")}}!function(e){if(e.srcset&&!u&&window.picturefill){var t=window.picturefill._;e[t.ns]&&e[t.ns].evaled||t.fillImg(e,{reselect:!0}),e[t.ns].curSrc||(e[t.ns].supported=!1,t.fillImg(e,{reselect:!0})),e.currentSrc=e[t.ns].curSrc||e.src}}(r.img),e.style.backgroundImage='url("'+(r.img.currentSrc||r.img.src).replace(/"/g,'\\"')+'")',e.style.backgroundPosition=t["object-position"]||"center",e.style.backgroundRepeat="no-repeat",e.style.backgroundOrigin="content-box",/scale-down/.test(t["object-fit"])?m(r.img,(function(){r.img.naturalWidth>e.width||r.img.naturalHeight>e.height?e.style.backgroundSize="contain":e.style.backgroundSize="auto"})):e.style.backgroundSize=t["object-fit"].replace("none","auto").replace("fill","100% 100%"),m(r.img,(function(t){p(e,t.naturalWidth,t.naturalHeight)}))}function g(e,t){var r=!d&&!e;if(t=t||{},e=e||"img",i&&!t.skipTest||!c)return!1;"img"===e?e=document.getElementsByTagName("img"):"string"==typeof e?e=document.querySelectorAll(e):"length"in e||(e=[e]);for(var o=0;o<e.length;o++)e[o][n]=e[o][n]||{skipTest:t.skipTest},v(e[o]);r&&(document.body.addEventListener("load",(function(e){"IMG"===e.target.tagName&&g(e.target,{skipTest:t.skipTest})}),!0),d=!0,e="img"),t.watchMQ&&window.addEventListener("resize",g.bind(null,e,{skipTest:t.skipTest}))}g.supportsObjectFit=a,g.supportsObjectPosition=i,function(){function e(e,t){return e[n]&&e[n].img&&("src"===t||"srcset"===t)?e[n].img:e}i||(HTMLImageElement.prototype.getAttribute=function(t){return l.call(e(this,t),t)},HTMLImageElement.prototype.setAttribute=function(t,r){return f.call(e(this,t),t,String(r))})}(),e.exports=g},7:function(e,t,r){"use strict";
-/*!
- * css-vars-ponyfill
- * v2.4.5
- * https://jhildenbiddle.github.io/css-vars-ponyfill/
- * (c) 2018-2021 John Hildenbiddle <http://hildenbiddle.com>
- * MIT license
- */
-function n(){return(n=Object.assign||function(e){for(var t=1;t<arguments.length;t++){var r=arguments[t];for(var n in r)Object.prototype.hasOwnProperty.call(r,n)&&(e[n]=r[n])}return e}).apply(this,arguments)}
-/*!
- * get-css-data
- * v2.0.2
- * https://github.com/jhildenbiddle/get-css-data
- * (c) 2018-2021 John Hildenbiddle <http://hildenbiddle.com>
- * MIT license
- */function o(e){var t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},r={mimeType:t.mimeType||null,onBeforeSend:t.onBeforeSend||Function.prototype,onSuccess:t.onSuccess||Function.prototype,onError:t.onError||Function.prototype,onComplete:t.onComplete||Function.prototype},n=Array.isArray(e)?e:[e],o=Array.apply(null,Array(n.length)).map((function(e){return null}));function s(e){var t="string"==typeof e,r=t&&"<"===e.trim().charAt(0);return t&&!r}function a(e,t){r.onError(e,n[t],t)}function i(e,t){var s=r.onSuccess(e,n[t],t);e=!1===s?"":s||e,o[t]=e,-1===o.indexOf(null)&&r.onComplete(o)}var c=document.createElement("a");n.forEach((function(e,t){if(c.setAttribute("href",e),c.href=String(c.href),Boolean(document.all&&!window.atob)&&c.host.split(":")[0]!==location.host.split(":")[0]){if(c.protocol===location.protocol){var n=new XDomainRequest;n.open("GET",e),n.timeout=0,n.onprogress=Function.prototype,n.ontimeout=Function.prototype,n.onload=function(){var e=n.responseText;s(e)?i(e,t):a(n,t)},n.onerror=function(e){a(n,t)},setTimeout((function(){n.send()}),0)}else console.warn("Internet Explorer 9 Cross-Origin (CORS) requests must use the same protocol (".concat(e,")")),a(null,t)}else{var o=new XMLHttpRequest;o.open("GET",e),r.mimeType&&o.overrideMimeType&&o.overrideMimeType(r.mimeType),r.onBeforeSend(o,e,t),o.onreadystatechange=function(){if(4===o.readyState){var e=o.responseText;o.status<400&&s(e)||0===o.status&&s(e)?i(e,t):a(o,t)}},o.send()}}))}
-/**
- * Gets CSS data from <style> and <link> nodes (including @imports), then
- * returns data in order processed by DOM. Allows specifying nodes to
- * include/exclude and filtering CSS data using RegEx.
- *
- * @preserve
- * @param {object}   [options] The options object
- * @param {object}   [options.rootElement=document] Root element to traverse for
- *                   <link> and <style> nodes.
- * @param {string}   [options.include] CSS selector matching <link> and <style>
- *                   nodes to include
- * @param {string}   [options.exclude] CSS selector matching <link> and <style>
- *                   nodes to exclude
- * @param {object}   [options.filter] Regular expression used to filter node CSS
- *                   data. Each block of CSS data is tested against the filter,
- *                   and only matching data is included.
- * @param {boolean}  [options.skipDisabled=true] Determines if disabled
- *                   stylesheets will be skipped while collecting CSS data.
- * @param {boolean}  [options.useCSSOM=false] Determines if CSS data will be
- *                   collected from a stylesheet's runtime values instead of its
- *                   text content. This is required to get accurate CSS data
- *                   when a stylesheet has been modified using the deleteRule()
- *                   or insertRule() methods because these modifications will
- *                   not be reflected in the stylesheet's text content.
- * @param {function} [options.onBeforeSend] Callback before XHR is sent. Passes
- *                   1) the XHR object, 2) source node reference, and 3) the
- *                   source URL as arguments.
- * @param {function} [options.onSuccess] Callback on each CSS node read. Passes
- *                   1) CSS text, 2) source node reference, and 3) the source
- *                   URL as arguments.
- * @param {function} [options.onError] Callback on each error. Passes 1) the XHR
- *                   object for inspection, 2) soure node reference, and 3) the
- *                   source URL that failed (either a <link> href or an @import)
- *                   as arguments
- * @param {function} [options.onComplete] Callback after all nodes have been
- *                   processed. Passes 1) concatenated CSS text, 2) an array of
- *                   CSS text in DOM order, and 3) an array of nodes in DOM
- *                   order as arguments.
- *
- * @example
- *
- *   getCssData({
- *     rootElement : document,
- *     include     : 'style,link[rel="stylesheet"]',
- *     exclude     : '[href="skip.css"]',
- *     filter      : /red/,
- *     skipDisabled: true,
- *     useCSSOM    : false,
- *     onBeforeSend(xhr, node, url) {
- *       // ...
- *     }
- *     onSuccess(cssText, node, url) {
- *       // ...
- *     }
- *     onError(xhr, node, url) {
- *       // ...
- *     },
- *     onComplete(cssText, cssArray, nodeArray) {
- *       // ...
- *     }
- *   });
- */function s(e){var t=/\/\*[\s\S]+?\*\//g,r=/(?:@import\s*)(?:url\(\s*)?(?:['"])([^'"]*)(?:['"])(?:\s*\))?(?:[^;]*;)/g,n={rootElement:e.rootElement||document,include:e.include||'style,link[rel="stylesheet"]',exclude:e.exclude||null,filter:e.filter||null,skipDisabled:!1!==e.skipDisabled,useCSSOM:e.useCSSOM||!1,onBeforeSend:e.onBeforeSend||Function.prototype,onSuccess:e.onSuccess||Function.prototype,onError:e.onError||Function.prototype,onComplete:e.onComplete||Function.prototype},s=Array.apply(null,n.rootElement.querySelectorAll(n.include)).filter((function(e){return t=e,r=n.exclude,!(t.matches||t.matchesSelector||t.webkitMatchesSelector||t.mozMatchesSelector||t.msMatchesSelector||t.oMatchesSelector).call(t,r);var t,r})),i=Array.apply(null,Array(s.length)).map((function(e){return null}));function c(){if(-1===i.indexOf(null)){i.reduce((function(e,t,r){return""===t&&e.push(r),e}),[]).reverse().forEach((function(e){return[s,i].forEach((function(t){return t.splice(e,1)}))}));var e=i.join("");n.onComplete(e,i,s)}}function u(e,t,r,s){var a=n.onSuccess(e,r,s);(function e(t,r,s,a){var i=arguments.length>4&&void 0!==arguments[4]?arguments[4]:[],c=arguments.length>5&&void 0!==arguments[5]?arguments[5]:[],u=l(t,s,c);u.rules.length?o(u.absoluteUrls,{onBeforeSend:function(e,t,o){n.onBeforeSend(e,r,t)},onSuccess:function(e,t,o){var s=n.onSuccess(e,r,t),a=l(e=!1===s?"":s||e,t,c);return a.rules.forEach((function(t,r){e=e.replace(t,a.absoluteRules[r])})),e},onError:function(n,o,l){i.push({xhr:n,url:o}),c.push(u.rules[l]),e(t,r,s,a,i,c)},onComplete:function(n){n.forEach((function(e,r){t=t.replace(u.rules[r],e)})),e(t,r,s,a,i,c)}}):a(t,i)})(e=void 0!==a&&!1===Boolean(a)?"":a||e,r,s,(function(e,o){null===i[t]&&(o.forEach((function(e){return n.onError(e.xhr,r,e.url)})),!n.filter||n.filter.test(e)?i[t]=e:i[t]="",c())}))}function l(e,n){var o=arguments.length>2&&void 0!==arguments[2]?arguments[2]:[],s={};return s.rules=(e.replace(t,"").match(r)||[]).filter((function(e){return-1===o.indexOf(e)})),s.urls=s.rules.map((function(e){return e.replace(r,"$1")})),s.absoluteUrls=s.urls.map((function(e){return a(e,n)})),s.absoluteRules=s.rules.map((function(e,t){var r=s.urls[t],o=a(s.absoluteUrls[t],n);return e.replace(r,o)})),s}s.length?s.forEach((function(e,t){var r=e.getAttribute("href"),s=e.getAttribute("rel"),l="link"===e.nodeName.toLowerCase()&&r&&s&&-1!==s.toLowerCase().indexOf("stylesheet"),f=!1!==n.skipDisabled&&e.disabled,d="style"===e.nodeName.toLowerCase();if(l&&!f)o(r,{mimeType:"text/css",onBeforeSend:function(t,r,o){n.onBeforeSend(t,e,r)},onSuccess:function(n,o,s){var i=a(r);u(n,t,e,i)},onError:function(r,o,s){i[t]="",n.onError(r,e,o),c()}});else if(d&&!f){var p=e.textContent;n.useCSSOM&&(p=Array.apply(null,e.sheet.cssRules).map((function(e){return e.cssText})).join("")),u(p,t,e,location.href)}else i[t]="",c()})):n.onComplete("",[])}function a(e,t){var r=document.implementation.createHTMLDocument(""),n=r.createElement("base"),o=r.createElement("a");return r.head.appendChild(n),r.body.appendChild(o),n.href=t||document.baseURI||(document.querySelector("base")||{}).href||location.href,o.href=e,o.href}r.r(t);var i=c;function c(e,t,r){e instanceof RegExp&&(e=u(e,r)),t instanceof RegExp&&(t=u(t,r));var n=l(e,t,r);return n&&{start:n[0],end:n[1],pre:r.slice(0,n[0]),body:r.slice(n[0]+e.length,n[1]),post:r.slice(n[1]+t.length)}}function u(e,t){var r=t.match(e);return r?r[0]:null}function l(e,t,r){var n,o,s,a,i,c=r.indexOf(e),u=r.indexOf(t,c+1),l=c;if(c>=0&&u>0){if(e===t)return[c,u];for(n=[],s=r.length;l>=0&&!i;)l==c?(n.push(l),c=r.indexOf(e,l+1)):1==n.length?i=[n.pop(),u]:((o=n.pop())<s&&(s=o,a=u),u=r.indexOf(t,l+1)),l=c<u&&c>=0?c:u;n.length&&(i=[s,a])}return i}function f(e){var t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},r={preserveStatic:!0,removeComments:!1},o=n({},r,t),s=[];function a(e){throw new Error("CSS parse error: ".concat(e))}function c(t){var r=t.exec(e);if(r)return e=e.slice(r[0].length),r}function u(){return c(/^{\s*/)}function l(){return c(/^}/)}function f(){c(/^\s*/)}function d(){if(f(),"/"===e[0]&&"*"===e[1]){for(var t=2;e[t]&&("*"!==e[t]||"/"!==e[t+1]);)t++;if(!e[t])return a("end of comment is missing");var r=e.slice(2,t);return e=e.slice(t+2),{type:"comment",comment:r}}}function p(){for(var e,t=[];e=d();)t.push(e);return o.removeComments?[]:t}function m(){for(f();"}"===e[0];)a("extra closing bracket");var t=c(/^(("(?:\\"|[^"])*"|'(?:\\'|[^'])*'|[^{])+)/);if(t)return t[0].trim().replace(/\/\*([^*]|[\r\n]|(\*+([^*/]|[\r\n])))*\*\/+/g,"").replace(/"(?:\\"|[^"])*"|'(?:\\'|[^'])*'/g,(function(e){return e.replace(/,/g,"‌")})).split(/\s*(?![^(]*\)),\s*/).map((function(e){return e.replace(/\u200C/g,",")}))}function v(){if("@"===e[0])return A();c(/^([;\s]*)+/);var t=/\/\*[^*]*\*+([^/*][^*]*\*+)*\//g,r=c(/^(\*?[-#/*\\\w]+(\[[0-9a-z_-]+\])?)\s*/);if(r){if(r=r[0].trim(),!c(/^:\s*/))return a("property missing ':'");var n=c(/^((?:\/\*.*?\*\/|'(?:\\'|.)*?'|"(?:\\"|.)*?"|\((\s*'(?:\\'|.)*?'|"(?:\\"|.)*?"|[^)]*?)\s*\)|[^};])+)/),o={type:"declaration",property:r.replace(t,""),value:n?n[0].replace(t,"").trim():""};return c(/^[;\s]*/),o}}function g(){if(!u())return a("missing '{'");for(var e,t=p();e=v();)t.push(e),t=t.concat(p());return l()?t:a("missing '}'")}function h(){f();for(var e,t=[];e=c(/^((\d+\.\d+|\.\d+|\d+)%?|[a-z]+)\s*/);)t.push(e[1]),c(/^,\s*/);if(t.length)return{type:"keyframe",values:t,declarations:g()}}function y(){var e=c(/^@([-\w]+)?keyframes\s*/);if(e){var t=e[1];if(!(e=c(/^([-\w]+)\s*/)))return a("@keyframes missing name");var r,n=e[1];if(!u())return a("@keyframes missing '{'");for(var o=p();r=h();)o.push(r),o=o.concat(p());return l()?{type:"keyframes",name:n,vendor:t,keyframes:o}:a("@keyframes missing '}'")}}function b(){if(c(/^@page */))return{type:"page",selectors:m()||[],declarations:g()}}function S(){var e=c(/@(top|bottom|left|right)-(left|center|right|top|middle|bottom)-?(corner)?\s*/);if(e)return{type:"page-margin-box",name:"".concat(e[1],"-").concat(e[2])+(e[3]?"-".concat(e[3]):""),declarations:g()}}function E(){if(c(/^@font-face\s*/))return{type:"font-face",declarations:g()}}function w(){var e=c(/^@supports *([^{]+)/);if(e)return{type:"supports",supports:e[1].trim(),rules:M()}}function C(){if(c(/^@host\s*/))return{type:"host",rules:M()}}function O(){var e=c(/^@media([^{]+)*/);if(e)return{type:"media",media:(e[1]||"").trim(),rules:M()}}function x(){var e=c(/^@custom-media\s+(--[^\s]+)\s*([^{;]+);/);if(e)return{type:"custom-media",name:e[1].trim(),media:e[2].trim()}}function j(){var e=c(/^@([-\w]+)?document *([^{]+)/);if(e)return{type:"document",document:e[2].trim(),vendor:e[1]?e[1].trim():null,rules:M()}}function k(){var e=c(/^@(import|charset|namespace)\s*([^;]+);/);if(e)return{type:e[1],name:e[2].trim()}}function A(){if(f(),"@"===e[0]){var t=k()||E()||O()||y()||w()||j()||x()||C()||b()||S();if(t&&!o.preserveStatic){var r=!1;if(t.declarations)r=t.declarations.some((function(e){return/var\(/.test(e.value)}));else r=(t.keyframes||t.rules||[]).some((function(e){return(e.declarations||[]).some((function(e){return/var\(/.test(e.value)}))}));return r?t:{}}return t}}function _(){if(!o.preserveStatic){var t=i("{","}",e);if(t){var r=/:(?:root|host)(?![.:#(])/.test(t.pre)&&/--\S*\s*:/.test(t.body),n=/var\(/.test(t.body);if(!r&&!n)return e=e.slice(t.end+1),{}}}var s=m()||[],c=o.preserveStatic?g():g().filter((function(e){var t=s.some((function(e){return/:(?:root|host)(?![.:#(])/.test(e)}))&&/^--\S/.test(e.property),r=/var\(/.test(e.value);return t||r}));return s.length||a("selector missing"),{type:"rule",selectors:s,declarations:c}}function M(t){if(!t&&!u())return a("missing '{'");for(var r,n=p();e.length&&(t||"}"!==e[0])&&(r=A()||_());)r.type&&n.push(r),n=n.concat(p());return t||l()?n:a("missing '}'")}return{type:"stylesheet",stylesheet:{rules:M(!0),errors:s}}}function d(e){var t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},r={parseHost:!1,store:{},onWarning:function(){}},o=n({},r,t),s=new RegExp(":".concat(o.parseHost?"host":"root","$"));return"string"==typeof e&&(e=f(e,o)),e.stylesheet.rules.forEach((function(e){"rule"===e.type&&e.selectors.some((function(e){return s.test(e)}))&&e.declarations.forEach((function(e,t){var r=e.property,n=e.value;r&&0===r.indexOf("--")&&(o.store[r]=n)}))})),o.store}function p(e){var t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:"",r=arguments.length>2?arguments[2]:void 0,n={charset:function(e){return"@charset "+e.name+";"},comment:function(e){return 0===e.comment.indexOf("__CSSVARSPONYFILL")?"/*"+e.comment+"*/":""},"custom-media":function(e){return"@custom-media "+e.name+" "+e.media+";"},declaration:function(e){return e.property+":"+e.value+";"},document:function(e){return"@"+(e.vendor||"")+"document "+e.document+"{"+o(e.rules)+"}"},"font-face":function(e){return"@font-face{"+o(e.declarations)+"}"},host:function(e){return"@host{"+o(e.rules)+"}"},import:function(e){return"@import "+e.name+";"},keyframe:function(e){return e.values.join(",")+"{"+o(e.declarations)+"}"},keyframes:function(e){return"@"+(e.vendor||"")+"keyframes "+e.name+"{"+o(e.keyframes)+"}"},media:function(e){return"@media "+e.media+"{"+o(e.rules)+"}"},namespace:function(e){return"@namespace "+e.name+";"},page:function(e){return"@page "+(e.selectors.length?e.selectors.join(", "):"")+"{"+o(e.declarations)+"}"},"page-margin-box":function(e){return"@"+e.name+"{"+o(e.declarations)+"}"},rule:function(e){var t=e.declarations;if(t.length)return e.selectors.join(",")+"{"+o(t)+"}"},supports:function(e){return"@supports "+e.supports+"{"+o(e.rules)+"}"}};function o(e){for(var o="",s=0;s<e.length;s++){var a=e[s];r&&r(a);var i=n[a.type](a);i&&(o+=i,i.length&&a.selectors&&(o+=t))}return o}return o(e.stylesheet.rules)}function m(e,t){e.rules.forEach((function(r){r.rules?m(r,t):r.keyframes?r.keyframes.forEach((function(e){"keyframe"===e.type&&t(e.declarations,r)})):r.declarations&&t(r.declarations,e)}))}c.range=l;function v(e){var t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},r={preserveStatic:!0,preserveVars:!1,variables:{},onWarning:function(){}},o=n({},r,t);return"string"==typeof e&&(e=f(e,o)),m(e.stylesheet,(function(e,t){for(var r=0;r<e.length;r++){var n=e[r],s=n.type,a=n.property,i=n.value;if("declaration"===s)if(o.preserveVars||!a||0!==a.indexOf("--")){if(-1!==i.indexOf("var(")){var c=h(i,o);c!==n.value&&(c=g(c),o.preserveVars?(e.splice(r,0,{type:s,property:a,value:c}),r++):n.value=c)}}else e.splice(r,1),r--}})),p(e)}function g(e){return(e.match(/calc\(([^)]+)\)/g)||[]).forEach((function(t){var r="calc".concat(t.split("calc").join(""));e=e.replace(t,r)})),e}function h(e){var t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},r=arguments.length>2?arguments[2]:void 0;if(-1===e.indexOf("var("))return e;var n=i("(",")",e);function o(e){var n=e.split(",")[0].replace(/[\s\n\t]/g,""),o=(e.match(/(?:\s*,\s*){1}(.*)?/)||[])[1],s=Object.prototype.hasOwnProperty.call(t.variables,n)?String(t.variables[n]):void 0,a=s||(o?String(o):void 0),i=r||e;return s||t.onWarning('variable "'.concat(n,'" is undefined')),a&&"undefined"!==a&&a.length>0?h(a,t,i):"var(".concat(i,")")}if(n){if("var"===n.pre.slice(-3)){var s=0===n.body.trim().length;return s?(t.onWarning("var() must contain a non-whitespace string"),e):n.pre.slice(0,-3)+o(n.body)+h(n.post,t)}return n.pre+"(".concat(h(n.body,t),")")+h(n.post,t)}return-1!==e.indexOf("var(")&&t.onWarning('missing closing ")" in the value "'.concat(e,'"')),e}var y="undefined"!=typeof window,b=y&&window.CSS&&window.CSS.supports&&window.CSS.supports("(--a: 0)"),S={group:0,job:0},E={rootElement:y?document:null,shadowDOM:!1,include:"style,link[rel=stylesheet]",exclude:"",variables:{},onlyLegacy:!0,preserveStatic:!0,preserveVars:!1,silent:!1,updateDOM:!0,updateURLs:!0,watch:null,onBeforeSend:function(){},onError:function(){},onWarning:function(){},onSuccess:function(){},onComplete:function(){},onFinally:function(){}},w={cssComments:/\/\*[\s\S]+?\*\//g,cssKeyframes:/@(?:-\w*-)?keyframes/,cssMediaQueries:/@media[^{]+\{([\s\S]+?})\s*}/g,cssUrls:/url\((?!['"]?(?:data|http|\/\/):)['"]?([^'")]*)['"]?\)/g,cssVarDeclRules:/(?::(?:root|host)(?![.:#(])[\s,]*[^{]*{\s*[^}]*})/g,cssVarDecls:/(?:[\s;]*)(-{2}\w[\w-]*)(?:\s*:\s*)([^;]*);/g,cssVarFunc:/var\(\s*--[\w-]/,cssVars:/(?:(?::(?:root|host)(?![.:#(])[\s,]*[^{]*{\s*[^;]*;*\s*)|(?:var\(\s*))(--[^:)]+)(?:\s*[:)])/},C={dom:{},job:{},user:{}},O=!1,x=null,j=0,k=null,A=!1;
-/**
- * Fetches, parses, and transforms CSS custom properties from specified
- * <style> and <link> elements into static values, then appends a new <style>
- * element with static values to the DOM to provide CSS custom property
- * compatibility for legacy browsers. Also provides a single interface for
- * live updates of runtime values in both modern and legacy browsers.
- *
- * @preserve
- * @param {object}   [options] Options object
- * @param {object}   [options.rootElement=document] Root element to traverse for
- *                   <link> and <style> nodes
- * @param {boolean}  [options.shadowDOM=false] Determines if shadow DOM <link>
- *                   and <style> nodes will be processed.
- * @param {string}   [options.include="style,link[rel=stylesheet]"] CSS selector
- *                   matching <link re="stylesheet"> and <style> nodes to
- *                   process
- * @param {string}   [options.exclude] CSS selector matching <link
- *                   rel="stylehseet"> and <style> nodes to exclude from those
- *                   matches by options.include
- * @param {object}   [options.variables] A map of custom property name/value
- *                   pairs. Property names can omit or include the leading
- *                   double-hyphen (—), and values specified will override
- *                   previous values
- * @param {boolean}  [options.onlyLegacy=true] Determines if the ponyfill will
- *                   only generate legacy-compatible CSS in browsers that lack
- *                   native support (i.e., legacy browsers)
- * @param {boolean}  [options.preserveStatic=true] Determines if CSS
- *                   declarations that do not reference a custom property will
- *                   be preserved in the transformed CSS
- * @param {boolean}  [options.preserveVars=false] Determines if CSS custom
- *                   property declarations will be preserved in the transformed
- *                   CSS
- * @param {boolean}  [options.silent=false] Determines if warning and error
- *                   messages will be displayed on the console
- * @param {boolean}  [options.updateDOM=true] Determines if the ponyfill will
- *                   update the DOM after processing CSS custom properties
- * @param {boolean}  [options.updateURLs=true] Determines if relative url()
- *                   paths will be converted to absolute urls in external CSS
- * @param {boolean}  [options.watch=false] Determines if a MutationObserver will
- *                   be created that will execute the ponyfill when a <link> or
- *                   <style> DOM mutation is observed
- * @param {function} [options.onBeforeSend] Callback before XHR is sent. Passes
- *                   1) the XHR object, 2) source node reference, and 3) the
- *                   source URL as arguments
- * @param {function} [options.onError] Callback after a CSS parsing error has
- *                   occurred or an XHR request has failed. Passes 1) an error
- *                   message, and 2) source node reference, 3) xhr, and 4 url as
- *                   arguments.
- * @param {function} [options.onWarning] Callback after each CSS parsing warning
- *                   has occurred. Passes 1) a warning message as an argument.
- * @param {function} [options.onSuccess] Callback after CSS data has been
- *                   collected from each node and before CSS custom properties
- *                   have been transformed. Allows modifying the CSS data before
- *                   it is transformed by returning any string value (or false
- *                   to skip). Passes 1) CSS text, 2) source node reference, and
- *                   3) the source URL as arguments.
- * @param {function} [options.onComplete] Callback after all CSS has been
- *                   processed, legacy-compatible CSS has been generated, and
- *                   (optionally) the DOM has been updated. Passes 1) a CSS
- *                   string with CSS variable values resolved, 2) an array of
- *                   output <style> node references that have been appended to
- *                   the DOM, 3) an object containing all custom properies names
- *                   and values, and 4) the ponyfill execution time in
- *                   milliseconds.
- * @param {function} [options.onFinally] Callback in modern and legacy browsers
- *                   after the ponyfill has finished all tasks. Passes 1) a
- *                   boolean indicating if the last ponyfill call resulted in a
- *                   style change, 2) a boolean indicating if the current
- *                   browser provides native support for CSS custom properties,
- *                   and 3) the ponyfill execution time in milliseconds.
- * @example
- *
- *   cssVars({
- *     rootElement   : document,
- *     shadowDOM     : false,
- *     include       : 'style,link[rel="stylesheet"]',
- *     exclude       : '',
- *     variables     : {},
- *     onlyLegacy    : true,
- *     preserveStatic: true,
- *     preserveVars  : false,
- *     silent        : false,
- *     updateDOM     : true,
- *     updateURLs    : true,
- *     watch         : false,
- *     onBeforeSend(xhr, node, url) {},
- *     onError(message, node, xhr, url) {},
- *     onWarning(message) {},
- *     onSuccess(cssText, node, url) {},
- *     onComplete(cssText, styleNode, cssVariables, benchmark) {},
- *     onFinally(hasChanged, hasNativeSupport, benchmark)
- *   });
- */
-function _(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{},t="cssVars(): ",r=n({},E,e);function o(e,n,o,s){!r.silent&&window.console&&console.error("".concat(t).concat(e,"\n"),n),r.onError(e,n,o,s)}function a(e){!r.silent&&window.console&&console.warn("".concat(t).concat(e)),r.onWarning(e)}function i(e){r.onFinally(Boolean(e),b,D()-r.__benchmark)}if(y){if(r.watch)return r.watch=E.watch,M(r),void _(r);if(!1===r.watch&&x&&(x.disconnect(),x=null),!r.__benchmark){if(O===r.rootElement)return void T(e);var c=[].slice.call(r.rootElement.querySelectorAll('[data-cssvars]:not([data-cssvars="out"])'));if(r.__benchmark=D(),r.exclude=[x?'[data-cssvars]:not([data-cssvars=""])':'[data-cssvars="out"]',"link[disabled]:not([data-cssvars])",r.exclude].filter((function(e){return e})).join(","),r.variables=R(r.variables),c.forEach((function(e){var t="style"===e.nodeName.toLowerCase()&&e.__cssVars.text,r=t&&e.textContent!==e.__cssVars.text;t&&r&&(e.sheet&&(e.sheet.disabled=!1),e.setAttribute("data-cssvars",""))})),!x){var u=[].slice.call(r.rootElement.querySelectorAll('[data-cssvars="out"]'));u.forEach((function(e){var t=e.getAttribute("data-cssvars-group");(t?r.rootElement.querySelector('[data-cssvars="src"][data-cssvars-group="'.concat(t,'"]')):null)||e.parentNode.removeChild(e)})),j&&c.length<j&&(j=c.length,C.dom={})}}if("loading"!==document.readyState)if(b&&r.onlyLegacy){var l=!1;if(r.updateDOM){var m=r.rootElement.host||(r.rootElement===document?document.documentElement:r.rootElement);Object.keys(r.variables).forEach((function(e){var t=r.variables[e];l=l||t!==getComputedStyle(m).getPropertyValue(e),m.style.setProperty(e,t)}))}i(l)}else!A&&(r.shadowDOM||r.rootElement.shadowRoot||r.rootElement.host)?s({rootElement:E.rootElement,include:E.include,exclude:r.exclude,skipDisabled:!1,onSuccess:function(e,t,r){return!((t.sheet||{}).disabled&&!t.__cssVars)&&((e=((e=e.replace(w.cssComments,"").replace(w.cssMediaQueries,"")).match(w.cssVarDeclRules)||[]).join(""))||!1)},onComplete:function(e,t,n){d(e,{store:C.dom,onWarning:a}),A=!0,_(r)}}):(O=r.rootElement,s({rootElement:r.rootElement,include:r.include,exclude:r.exclude,skipDisabled:!1,onBeforeSend:r.onBeforeSend,onError:function(e,t,r){var n=e.responseURL||N(r,location.href),s=e.statusText?"(".concat(e.statusText,")"):"Unspecified Error"+(0===e.status?" (possibly CORS related)":"");o("CSS XHR Error: ".concat(n," ").concat(e.status," ").concat(s),t,e,n)},onSuccess:function(e,t,n){if((t.sheet||{}).disabled&&!t.__cssVars)return!1;var o="link"===t.nodeName.toLowerCase(),s="style"===t.nodeName.toLowerCase()&&e!==t.textContent,a=r.onSuccess(e,t,n);return e=void 0!==a&&!1===Boolean(a)?"":a||e,r.updateURLs&&(o||s)&&(e=V(e,n)),e},onComplete:function(e,t){var s=arguments.length>2&&void 0!==arguments[2]?arguments[2]:[],c=n({},C.dom,C.user);if(C.job={},s.forEach((function(e,n){var s=t[n];if(e.__cssVars=e.__cssVars||{},e.__cssVars.text=s,w.cssVars.test(s))try{var i=f(s,{preserveStatic:r.preserveStatic,removeComments:!0});d(i,{parseHost:Boolean(r.rootElement.host),store:C.dom,onWarning:a}),e.__cssVars.tree=i}catch(t){o(t.message,e)}})),n(C.job,C.dom),r.updateDOM?(n(C.user,r.variables),n(C.job,C.user)):(n(C.job,C.user,r.variables),n(c,r.variables)),S.job>0&&Boolean(Object.keys(C.job).length>Object.keys(c).length||Boolean(Object.keys(c).length&&Object.keys(C.job).some((function(e){return C.job[e]!==c[e]})))))B(r.rootElement),_(r);else{var u=[],l=[],m=!1;if(r.updateDOM&&S.job++,s.forEach((function(e,s){var i=!e.__cssVars.tree;if(e.__cssVars.tree)try{v(e.__cssVars.tree,n({},r,{variables:C.job,onWarning:a}));var c=p(e.__cssVars.tree);if(r.updateDOM){var f=t[s],d=w.cssVarFunc.test(f);if(e.getAttribute("data-cssvars")||e.setAttribute("data-cssvars","src"),c.length&&d){var g=e.getAttribute("data-cssvars-group")||++S.group,h=c.replace(/\s/g,""),y=r.rootElement.querySelector('[data-cssvars="out"][data-cssvars-group="'.concat(g,'"]'))||document.createElement("style");m=m||w.cssKeyframes.test(c),r.preserveStatic&&e.sheet&&(e.sheet.disabled=!0),y.hasAttribute("data-cssvars")||y.setAttribute("data-cssvars","out"),h===e.textContent.replace(/\s/g,"")?(i=!0,y&&y.parentNode&&(e.removeAttribute("data-cssvars-group"),y.parentNode.removeChild(y))):h!==y.textContent.replace(/\s/g,"")&&([e,y].forEach((function(e){e.setAttribute("data-cssvars-job",S.job),e.setAttribute("data-cssvars-group",g)})),y.textContent=c,u.push(c),l.push(y),y.parentNode||e.parentNode.insertBefore(y,e.nextSibling))}}else e.textContent.replace(/\s/g,"")!==c&&u.push(c)}catch(t){o(t.message,e)}i&&e.setAttribute("data-cssvars","skip"),e.hasAttribute("data-cssvars-job")||e.setAttribute("data-cssvars-job",S.job)})),j=r.rootElement.querySelectorAll('[data-cssvars]:not([data-cssvars="out"])').length,r.shadowDOM)for(var g,h=[].concat(r.rootElement).concat([].slice.call(r.rootElement.querySelectorAll("*"))),y=0;g=h[y];++y)if(g.shadowRoot&&g.shadowRoot.querySelector("style")){var b=n({},r,{rootElement:g.shadowRoot});_(b)}r.updateDOM&&m&&L(r.rootElement),O=!1,r.onComplete(u.join(""),l,JSON.parse(JSON.stringify(C.job)),D()-r.__benchmark),i(l.length)}}}));else document.addEventListener("DOMContentLoaded",(function t(r){_(e),document.removeEventListener("DOMContentLoaded",t)}))}}function M(e){function t(e){var t=r(e)&&e.hasAttribute("disabled"),n=(e.sheet||{}).disabled;return t||n}function r(e){return"link"===e.nodeName.toLowerCase()&&-1!==(e.getAttribute("rel")||"").indexOf("stylesheet")}function n(e){return"style"===e.nodeName.toLowerCase()}window.MutationObserver&&(x&&(x.disconnect(),x=null),(x=new MutationObserver((function(o){o.some((function(o){return function(n){var o=!1;if("attributes"===n.type&&r(n.target)&&!t(n.target)){var s="disabled"===n.attributeName,a="href"===n.attributeName,i="skip"===n.target.getAttribute("data-cssvars"),c="src"===n.target.getAttribute("data-cssvars");s?o=!i&&!c:a&&(i?n.target.setAttribute("data-cssvars",""):c&&B(e.rootElement,!0),o=!0)}return o}(o)||function(e){var t=!1;if("childList"===e.type){var r=n(e.target),o="out"===e.target.getAttribute("data-cssvars");t=r&&!o}return t}(o)||function(e){var o=!1;return"childList"===e.type&&(o=[].slice.call(e.addedNodes).some((function(e){var o=1===e.nodeType&&e.hasAttribute("data-cssvars"),s=n(e)&&w.cssVars.test(e.textContent);return!o&&(r(e)||s)&&!t(e)}))),o}(o)||function(t){var r=!1;return"childList"===t.type&&(r=[].slice.call(t.removedNodes).some((function(t){var r=1===t.nodeType,n=r&&"out"===t.getAttribute("data-cssvars"),o=r&&"src"===t.getAttribute("data-cssvars"),s=o;if(o||n){var a=t.getAttribute("data-cssvars-group"),i=e.rootElement.querySelector('[data-cssvars-group="'.concat(a,'"]'));o&&B(e.rootElement,!0),i&&i.parentNode.removeChild(i)}return s}))),r}(o)}))&&_(e)}))).observe(document.documentElement,{attributes:!0,attributeFilter:["disabled","href"],childList:!0,subtree:!0}))}function T(e){var t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:100;clearTimeout(k),k=setTimeout((function(){e.__benchmark=null,_(e)}),t)}function L(e){var t=["animation-name","-moz-animation-name","-webkit-animation-name"].filter((function(e){return getComputedStyle(document.body)[e]}))[0];if(t){for(var r=e.getElementsByTagName("*"),n=[],o=0,s=r.length;o<s;o++){var a=r[o];"none"!==getComputedStyle(a)[t]&&(a.style[t]+="__CSSVARSPONYFILL-KEYFRAMES__",n.push(a))}document.body.offsetHeight;for(var i=0,c=n.length;i<c;i++){var u=n[i].style;u[t]=u[t].replace("__CSSVARSPONYFILL-KEYFRAMES__","")}}}function V(e,t){return(e.replace(w.cssComments,"").match(w.cssUrls)||[]).forEach((function(r){var n=r.replace(w.cssUrls,"$1"),o=N(n,t);e=e.replace(r,r.replace(n,o))})),e}function R(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{},t=/^-{2}/;return Object.keys(e).reduce((function(r,n){return r[t.test(n)?n:"--".concat(n.replace(/^-+/,""))]=e[n],r}),{})}function N(e){var t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:location.href,r=document.implementation.createHTMLDocument(""),n=r.createElement("base"),o=r.createElement("a");return r.head.appendChild(n),r.body.appendChild(o),n.href=t,o.href=e,o.href}function D(){return y&&(window.performance||{}).now?window.performance.now():(new Date).getTime()}function B(e){var t=arguments.length>1&&void 0!==arguments[1]&&arguments[1],r=[].slice.call(e.querySelectorAll('[data-cssvars="skip"],[data-cssvars="src"]'));r.forEach((function(e){return e.setAttribute("data-cssvars","")})),t&&(C.dom={})}_.reset=function(){for(var e in S.job=0,S.group=0,O=!1,x&&(x.disconnect(),x=null),j=0,k=null,A=!1,C)C[e]={}};var F=_,P=r(1),q=r.n(P);document.addEventListener("DOMContentLoaded",(function(){F(),q()(null,{watchMQ:!0})}))}});
+/*! For license information please see polyfill.js.LICENSE.txt */
+!function(){"use strict";var e={269:function(e){var t="bfred-it:object-fit-images",r=/(object-fit|object-position)\s*:\s*([-.\w\s%]+)/g,n="undefined"==typeof Image?{style:{"object-position":1}}:new Image,o="object-fit"in n.style,s="object-position"in n.style,a="background-size"in n.style,i="string"==typeof n.currentSrc,c=n.getAttribute,u=n.setAttribute,l=!1;function f(e,t,r){var n="data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='"+(t||1)+"' height='"+(r||0)+"'%3E%3C/svg%3E";c.call(e,"src")!==n&&u.call(e,"src",n)}function d(e,t){e.naturalWidth?t(e):setTimeout(d,100,e,t)}function p(e){var n=function(e){for(var t,n=getComputedStyle(e).fontFamily,o={};null!==(t=r.exec(n));)o[t[1]]=t[2];return o}(e),s=e[t];if(n["object-fit"]=n["object-fit"]||"fill",!s.img){if("fill"===n["object-fit"])return;if(!s.skipTest&&o&&!n["object-position"])return}if(!s.img){s.img=new Image(e.width,e.height),s.img.srcset=c.call(e,"data-ofi-srcset")||e.srcset,s.img.src=c.call(e,"data-ofi-src")||e.src,u.call(e,"data-ofi-src",e.src),e.srcset&&u.call(e,"data-ofi-srcset",e.srcset),f(e,e.naturalWidth||e.width,e.naturalHeight||e.height),e.srcset&&(e.srcset="");try{!function(e){var r={get:function(r){return e[t].img[r||"src"]},set:function(r,n){return e[t].img[n||"src"]=r,u.call(e,"data-ofi-"+n,r),p(e),r}};Object.defineProperty(e,"src",r),Object.defineProperty(e,"currentSrc",{get:function(){return r.get("currentSrc")}}),Object.defineProperty(e,"srcset",{get:function(){return r.get("srcset")},set:function(e){return r.set(e,"srcset")}})}(e)}catch(e){window.console&&console.warn("https://bit.ly/ofi-old-browser")}}!function(e){if(e.srcset&&!i&&window.picturefill){var t=window.picturefill._;e[t.ns]&&e[t.ns].evaled||t.fillImg(e,{reselect:!0}),e[t.ns].curSrc||(e[t.ns].supported=!1,t.fillImg(e,{reselect:!0})),e.currentSrc=e[t.ns].curSrc||e.src}}(s.img),e.style.backgroundImage='url("'+(s.img.currentSrc||s.img.src).replace(/"/g,'\\"')+'")',e.style.backgroundPosition=n["object-position"]||"center",e.style.backgroundRepeat="no-repeat",e.style.backgroundOrigin="content-box",/scale-down/.test(n["object-fit"])?d(s.img,(function(){s.img.naturalWidth>e.width||s.img.naturalHeight>e.height?e.style.backgroundSize="contain":e.style.backgroundSize="auto"})):e.style.backgroundSize=n["object-fit"].replace("none","auto").replace("fill","100% 100%"),d(s.img,(function(t){f(e,t.naturalWidth,t.naturalHeight)}))}function m(e,r){var n=!l&&!e;if(r=r||{},e=e||"img",s&&!r.skipTest||!a)return!1;"img"===e?e=document.getElementsByTagName("img"):"string"==typeof e?e=document.querySelectorAll(e):"length"in e||(e=[e]);for(var o=0;o<e.length;o++)e[o][t]=e[o][t]||{skipTest:r.skipTest},p(e[o]);n&&(document.body.addEventListener("load",(function(e){"IMG"===e.target.tagName&&m(e.target,{skipTest:r.skipTest})}),!0),l=!0,e="img"),r.watchMQ&&window.addEventListener("resize",m.bind(null,e,{skipTest:r.skipTest}))}m.supportsObjectFit=o,m.supportsObjectPosition=s,function(){function e(e,r){return e[t]&&e[t].img&&("src"===r||"srcset"===r)?e[t].img:e}s||(HTMLImageElement.prototype.getAttribute=function(t){return c.call(e(this,t),t)},HTMLImageElement.prototype.setAttribute=function(t,r){return u.call(e(this,t),t,String(r))})}(),e.exports=m}},t={};function r(n){var o=t[n];if(void 0!==o)return o.exports;var s=t[n]={exports:{}};return e[n](s,s.exports,r),s.exports}r.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return r.d(t,{a:t}),t},r.d=function(e,t){for(var n in t)r.o(t,n)&&!r.o(e,n)&&Object.defineProperty(e,n,{enumerable:!0,get:t[n]})},r.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},function(){function e(){return(e=Object.assign||function(e){for(var t=1;t<arguments.length;t++){var r=arguments[t];for(var n in r)Object.prototype.hasOwnProperty.call(r,n)&&(e[n]=r[n])}return e}).apply(this,arguments)}function t(e){var t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},r={mimeType:t.mimeType||null,onBeforeSend:t.onBeforeSend||Function.prototype,onSuccess:t.onSuccess||Function.prototype,onError:t.onError||Function.prototype,onComplete:t.onComplete||Function.prototype},n=Array.isArray(e)?e:[e],o=Array.apply(null,Array(n.length)).map((function(e){return null}));function s(e){var t="string"==typeof e,r=t&&"<"===e.trim().charAt(0);return t&&!r}function a(e,t){r.onError(e,n[t],t)}function i(e,t){var s=r.onSuccess(e,n[t],t);e=!1===s?"":s||e,o[t]=e,-1===o.indexOf(null)&&r.onComplete(o)}var c=document.createElement("a");n.forEach((function(e,t){if(c.setAttribute("href",e),c.href=String(c.href),Boolean(document.all&&!window.atob)&&c.host.split(":")[0]!==location.host.split(":")[0])if(c.protocol===location.protocol){var n=new XDomainRequest;n.open("GET",e),n.timeout=0,n.onprogress=Function.prototype,n.ontimeout=Function.prototype,n.onload=function(){var e=n.responseText;s(e)?i(e,t):a(n,t)},n.onerror=function(e){a(n,t)},setTimeout((function(){n.send()}),0)}else console.warn("Internet Explorer 9 Cross-Origin (CORS) requests must use the same protocol (".concat(e,")")),a(null,t);else{var o=new XMLHttpRequest;o.open("GET",e),r.mimeType&&o.overrideMimeType&&o.overrideMimeType(r.mimeType),r.onBeforeSend(o,e,t),o.onreadystatechange=function(){if(4===o.readyState){var e=o.responseText;o.status<400&&s(e)||0===o.status&&s(e)?i(e,t):a(o,t)}},o.send()}}))}function n(e){var r=/\/\*[\s\S]+?\*\//g,n=/(?:@import\s*)(?:url\(\s*)?(?:['"])([^'"]*)(?:['"])(?:\s*\))?(?:[^;]*;)/g,s={rootElement:e.rootElement||document,include:e.include||'style,link[rel="stylesheet"]',exclude:e.exclude||null,filter:e.filter||null,skipDisabled:!1!==e.skipDisabled,useCSSOM:e.useCSSOM||!1,onBeforeSend:e.onBeforeSend||Function.prototype,onSuccess:e.onSuccess||Function.prototype,onError:e.onError||Function.prototype,onComplete:e.onComplete||Function.prototype},a=Array.apply(null,s.rootElement.querySelectorAll(s.include)).filter((function(e){return t=e,r=s.exclude,!(t.matches||t.matchesSelector||t.webkitMatchesSelector||t.mozMatchesSelector||t.msMatchesSelector||t.oMatchesSelector).call(t,r);var t,r})),i=Array.apply(null,Array(a.length)).map((function(e){return null}));function c(){if(-1===i.indexOf(null)){i.reduce((function(e,t,r){return""===t&&e.push(r),e}),[]).reverse().forEach((function(e){return[a,i].forEach((function(t){return t.splice(e,1)}))}));var e=i.join("");s.onComplete(e,i,a)}}function u(e,t,r,n){var o=s.onSuccess(e,r,n);f(e=void 0!==o&&!1===Boolean(o)?"":o||e,r,n,(function(e,n){null===i[t]&&(n.forEach((function(e){return s.onError(e.xhr,r,e.url)})),!s.filter||s.filter.test(e)?i[t]=e:i[t]="",c())}))}function l(e,t){var s=arguments.length>2&&void 0!==arguments[2]?arguments[2]:[],a={};return a.rules=(e.replace(r,"").match(n)||[]).filter((function(e){return-1===s.indexOf(e)})),a.urls=a.rules.map((function(e){return e.replace(n,"$1")})),a.absoluteUrls=a.urls.map((function(e){return o(e,t)})),a.absoluteRules=a.rules.map((function(e,r){var n=a.urls[r],s=o(a.absoluteUrls[r],t);return e.replace(n,s)})),a}function f(e,r,n,o){var a=arguments.length>4&&void 0!==arguments[4]?arguments[4]:[],i=arguments.length>5&&void 0!==arguments[5]?arguments[5]:[],c=l(e,n,i);c.rules.length?t(c.absoluteUrls,{onBeforeSend:function(e,t,n){s.onBeforeSend(e,r,t)},onSuccess:function(e,t,n){var o=s.onSuccess(e,r,t),a=l(e=!1===o?"":o||e,t,i);return a.rules.forEach((function(t,r){e=e.replace(t,a.absoluteRules[r])})),e},onError:function(t,s,u){a.push({xhr:t,url:s}),i.push(c.rules[u]),f(e,r,n,o,a,i)},onComplete:function(t){t.forEach((function(t,r){e=e.replace(c.rules[r],t)})),f(e,r,n,o,a,i)}}):o(e,a)}a.length?a.forEach((function(e,r){var n=e.getAttribute("href"),a=e.getAttribute("rel"),l="link"===e.nodeName.toLowerCase()&&n&&a&&-1!==a.toLowerCase().indexOf("stylesheet"),f=!1!==s.skipDisabled&&e.disabled,d="style"===e.nodeName.toLowerCase();if(l&&!f)t(n,{mimeType:"text/css",onBeforeSend:function(t,r,n){s.onBeforeSend(t,e,r)},onSuccess:function(t,s,a){var i=o(n);u(t,r,e,i)},onError:function(t,n,o){i[r]="",s.onError(t,e,n),c()}});else if(d&&!f){var p=e.textContent;s.useCSSOM&&(p=Array.apply(null,e.sheet.cssRules).map((function(e){return e.cssText})).join("")),u(p,r,e,location.href)}else i[r]="",c()})):s.onComplete("",[])}function o(e,t){var r=document.implementation.createHTMLDocument(""),n=r.createElement("base"),o=r.createElement("a");return r.head.appendChild(n),r.body.appendChild(o),n.href=t||document.baseURI||(document.querySelector("base")||{}).href||location.href,o.href=e,o.href}var s=a;function a(e,t,r){e instanceof RegExp&&(e=i(e,r)),t instanceof RegExp&&(t=i(t,r));var n=c(e,t,r);return n&&{start:n[0],end:n[1],pre:r.slice(0,n[0]),body:r.slice(n[0]+e.length,n[1]),post:r.slice(n[1]+t.length)}}function i(e,t){var r=t.match(e);return r?r[0]:null}function c(e,t,r){var n,o,s,a,i,c=r.indexOf(e),u=r.indexOf(t,c+1),l=c;if(c>=0&&u>0){if(e===t)return[c,u];for(n=[],s=r.length;l>=0&&!i;)l==c?(n.push(l),c=r.indexOf(e,l+1)):1==n.length?i=[n.pop(),u]:((o=n.pop())<s&&(s=o,a=u),u=r.indexOf(t,l+1)),l=c<u&&c>=0?c:u;n.length&&(i=[s,a])}return i}function u(t){var r=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},n={preserveStatic:!0,removeComments:!1},o=e({},n,r),a=[];function i(e){throw new Error("CSS parse error: ".concat(e))}function c(e){var r=e.exec(t);if(r)return t=t.slice(r[0].length),r}function u(){return c(/^{\s*/)}function l(){return c(/^}/)}function f(){c(/^\s*/)}function d(){if(f(),"/"===t[0]&&"*"===t[1]){for(var e=2;t[e]&&("*"!==t[e]||"/"!==t[e+1]);)e++;if(!t[e])return i("end of comment is missing");var r=t.slice(2,e);return t=t.slice(e+2),{type:"comment",comment:r}}}function p(){for(var e,t=[];e=d();)t.push(e);return o.removeComments?[]:t}function m(){for(f();"}"===t[0];)i("extra closing bracket");var e=c(/^(("(?:\\"|[^"])*"|'(?:\\'|[^'])*'|[^{])+)/);if(e){var r,n=e[0].trim();/\/\*/.test(n)&&(n=n.replace(/\/\*([^*]|[\r\n]|(\*+([^*/]|[\r\n])))*\*\/+/g,""));var o=/["']\w*,\w*["']/.test(n);return o&&(n=n.replace(/"(?:\\"|[^"])*"|'(?:\\'|[^'])*'/g,(function(e){return e.replace(/,/g,"‌")}))),r=/,/.test(n)?n.split(/\s*(?![^(]*\)),\s*/):[n],o&&(r=r.map((function(e){return e.replace(/\u200C/g,",")}))),r}}function v(){if("@"===t[0])return A();c(/^([;\s]*)+/);var e=/\/\*[^*]*\*+([^/*][^*]*\*+)*\//g,r=c(/^(\*?[-#/*\\\w]+(\[[0-9a-z_-]+\])?)\s*/);if(r){if(r=r[0].trim(),!c(/^:\s*/))return i("property missing ':'");var n=c(/^((?:\/\*.*?\*\/|'(?:\\'|.)*?'|"(?:\\"|.)*?"|\((\s*'(?:\\'|.)*?'|"(?:\\"|.)*?"|[^)]*?)\s*\)|[^};])+)/),o={type:"declaration",property:r.replace(e,""),value:n?n[0].replace(e,"").trim():""};return c(/^[;\s]*/),o}}function g(){if(!u())return i("missing '{'");for(var e,t=p();e=v();)t.push(e),t=t.concat(p());return l()?t:i("missing '}'")}function h(){f();for(var e,t=[];e=c(/^((\d+\.\d+|\.\d+|\d+)%?|[a-z]+)\s*/);)t.push(e[1]),c(/^,\s*/);if(t.length)return{type:"keyframe",values:t,declarations:g()}}function y(){var e=c(/^@([-\w]+)?keyframes\s*/);if(e){var t=e[1];if(!(e=c(/^([-\w]+)\s*/)))return i("@keyframes missing name");var r,n=e[1];if(!u())return i("@keyframes missing '{'");for(var o=p();r=h();)o.push(r),o=o.concat(p());return l()?{type:"keyframes",name:n,vendor:t,keyframes:o}:i("@keyframes missing '}'")}}function b(){if(c(/^@page */))return{type:"page",selectors:m()||[],declarations:g()}}function S(){var e=c(/@(top|bottom|left|right)-(left|center|right|top|middle|bottom)-?(corner)?\s*/);if(e)return{type:"page-margin-box",name:"".concat(e[1],"-").concat(e[2])+(e[3]?"-".concat(e[3]):""),declarations:g()}}function w(){if(c(/^@font-face\s*/))return{type:"font-face",declarations:g()}}function E(){var e=c(/^@supports *([^{]+)/);if(e)return{type:"supports",supports:e[1].trim(),rules:M()}}function C(){if(c(/^@host\s*/))return{type:"host",rules:M()}}function x(){var e=c(/^@media([^{]+)*/);if(e)return{type:"media",media:(e[1]||"").trim(),rules:M()}}function k(){var e=c(/^@custom-media\s+(--[^\s]+)\s*([^{;]+);/);if(e)return{type:"custom-media",name:e[1].trim(),media:e[2].trim()}}function O(){var e=c(/^@([-\w]+)?document *([^{]+)/);if(e)return{type:"document",document:e[2].trim(),vendor:e[1]?e[1].trim():null,rules:M()}}function j(){var e=c(/^@(import|charset|namespace)\s*([^;]+);/);if(e)return{type:e[1],name:e[2].trim()}}function A(){if(f(),"@"===t[0]){var e=j()||w()||x()||y()||E()||O()||k()||C()||b()||S();if(e&&!o.preserveStatic){return(e.declarations?e.declarations.some((function(e){return/var\(/.test(e.value)})):(e.keyframes||e.rules||[]).some((function(e){return(e.declarations||[]).some((function(e){return/var\(/.test(e.value)}))})))?e:{}}return e}}function _(){if(!o.preserveStatic){var e=s("{","}",t);if(e){var r=/:(?:root|host)(?![.:#(])/.test(e.pre)&&/--\S*\s*:/.test(e.body),n=/var\(/.test(e.body);if(!r&&!n)return t=t.slice(e.end+1),{}}}var a=m()||[],c=o.preserveStatic?g():g().filter((function(e){var t=a.some((function(e){return/:(?:root|host)(?![.:#(])/.test(e)}))&&/^--\S/.test(e.property),r=/var\(/.test(e.value);return t||r}));return a.length||i("selector missing"),{type:"rule",selectors:a,declarations:c}}function M(e){if(!e&&!u())return i("missing '{'");for(var r,n=p();t.length&&(e||"}"!==t[0])&&(r=A()||_());)r.type&&n.push(r),n=n.concat(p());return e||l()?n:i("missing '}'")}return{type:"stylesheet",stylesheet:{rules:M(!0),errors:a}}}function l(t){var r=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},n={parseHost:!1,store:{},onWarning:function(){}},o=e({},n,r),s=new RegExp(":".concat(o.parseHost?"host":"root","$"));return"string"==typeof t&&(t=u(t,o)),t.stylesheet.rules.forEach((function(e){"rule"===e.type&&e.selectors.some((function(e){return s.test(e)}))&&e.declarations.forEach((function(e,t){var r=e.property,n=e.value;r&&0===r.indexOf("--")&&(o.store[r]=n)}))})),o.store}function f(e){var t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:"",r=arguments.length>2?arguments[2]:void 0,n={charset:function(e){return"@charset "+e.name+";"},comment:function(e){return 0===e.comment.indexOf("__CSSVARSPONYFILL")?"/*"+e.comment+"*/":""},"custom-media":function(e){return"@custom-media "+e.name+" "+e.media+";"},declaration:function(e){return e.property+":"+e.value+";"},document:function(e){return"@"+(e.vendor||"")+"document "+e.document+"{"+o(e.rules)+"}"},"font-face":function(e){return"@font-face{"+o(e.declarations)+"}"},host:function(e){return"@host{"+o(e.rules)+"}"},import:function(e){return"@import "+e.name+";"},keyframe:function(e){return e.values.join(",")+"{"+o(e.declarations)+"}"},keyframes:function(e){return"@"+(e.vendor||"")+"keyframes "+e.name+"{"+o(e.keyframes)+"}"},media:function(e){return"@media "+e.media+"{"+o(e.rules)+"}"},namespace:function(e){return"@namespace "+e.name+";"},page:function(e){return"@page "+(e.selectors.length?e.selectors.join(", "):"")+"{"+o(e.declarations)+"}"},"page-margin-box":function(e){return"@"+e.name+"{"+o(e.declarations)+"}"},rule:function(e){var t=e.declarations;if(t.length)return e.selectors.join(",")+"{"+o(t)+"}"},supports:function(e){return"@supports "+e.supports+"{"+o(e.rules)+"}"}};function o(e){for(var o="",s=0;s<e.length;s++){var a=e[s];r&&r(a);var i=n[a.type](a);i&&(o+=i,i.length&&a.selectors&&(o+=t))}return o}return o(e.stylesheet.rules)}function d(e,t){e.rules.forEach((function(r){r.rules?d(r,t):r.keyframes?r.keyframes.forEach((function(e){"keyframe"===e.type&&t(e.declarations,r)})):r.declarations&&t(r.declarations,e)}))}function p(t){var r=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},n={preserveStatic:!0,preserveVars:!1,variables:{},onWarning:function(){}},o=e({},n,r);return"string"==typeof t&&(t=u(t,o)),d(t.stylesheet,(function(e,t){for(var r=0;r<e.length;r++){var n=e[r],s=n.type,a=n.property,i=n.value;if("declaration"===s)if(o.preserveVars||!a||0!==a.indexOf("--")){if(-1!==i.indexOf("var(")){var c=v(i,o);c!==n.value&&(c=m(c),o.preserveVars?(e.splice(r,0,{type:s,property:a,value:c}),r++):n.value=c)}}else e.splice(r,1),r--}})),f(t)}function m(e){return(e.match(/calc\(([^)]+)\)/g)||[]).forEach((function(t){var r="calc".concat(t.split("calc").join(""));e=e.replace(t,r)})),e}function v(e){var t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},r=arguments.length>2?arguments[2]:void 0;if(-1===e.indexOf("var("))return e;var n=s("(",")",e);function o(e){var n=e.split(",")[0].replace(/[\s\n\t]/g,""),o=(e.match(/(?:\s*,\s*){1}(.*)?/)||[])[1],s=Object.prototype.hasOwnProperty.call(t.variables,n)?String(t.variables[n]):void 0,a=s||(o?String(o):void 0),i=r||e;return s||t.onWarning('variable "'.concat(n,'" is undefined')),a&&"undefined"!==a&&a.length>0?v(a,t,i):"var(".concat(i,")")}if(n){if("var"===n.pre.slice(-3)){var a=0===n.body.trim().length;return a?(t.onWarning("var() must contain a non-whitespace string"),e):n.pre.slice(0,-3)+o(n.body)+v(n.post,t)}return n.pre+"(".concat(v(n.body,t),")")+v(n.post,t)}return-1!==e.indexOf("var(")&&t.onWarning('missing closing ")" in the value "'.concat(e,'"')),e}a.range=c;var g="undefined"!=typeof window,h=g&&window.CSS&&window.CSS.supports&&window.CSS.supports("(--a: 0)"),y={group:0,job:0},b={rootElement:g?document:null,shadowDOM:!1,include:"style,link[rel=stylesheet]",exclude:"",variables:{},onlyLegacy:!0,preserveStatic:!0,preserveVars:!1,silent:!1,updateDOM:!0,updateURLs:!0,watch:null,onBeforeSend:function(){},onError:function(){},onWarning:function(){},onSuccess:function(){},onComplete:function(){},onFinally:function(){}},S={cssComments:/\/\*[\s\S]+?\*\//g,cssKeyframes:/@(?:-\w*-)?keyframes/,cssMediaQueries:/@media[^{]+\{([\s\S]+?})\s*}/g,cssUrls:/url\((?!['"]?(?:data|http|\/\/):)['"]?([^'")]*)['"]?\)/g,cssVarDeclRules:/(?::(?:root|host)(?![.:#(])[\s,]*[^{]*{\s*[^}]*})/g,cssVarDecls:/(?:[\s;]*)(-{2}\w[\w-]*)(?:\s*:\s*)([^;]*);/g,cssVarFunc:/var\(\s*--[\w-]/,cssVars:/(?:(?::(?:root|host)(?![.:#(])[\s,]*[^{]*{\s*[^;]*;*\s*)|(?:var\(\s*))(--[^:)]+)(?:\s*[:)])/},w={dom:{},job:{},user:{}},E=!1,C=null,x=0,k=null,O=!1;function j(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{},r="cssVars(): ",o=e({},b,t);function s(e,t,n,s){!o.silent&&window.console&&console.error("".concat(r).concat(e,"\n"),t),o.onError(e,t,n,s)}function a(e){!o.silent&&window.console&&console.warn("".concat(r).concat(e)),o.onWarning(e)}function i(e){o.onFinally(Boolean(e),h,R()-o.__benchmark)}if(g){if(o.watch)return o.watch=b.watch,A(o),void j(o);if(!1===o.watch&&C&&(C.disconnect(),C=null),!o.__benchmark){if(E===o.rootElement)return void _(t);var c=[].slice.call(o.rootElement.querySelectorAll('[data-cssvars]:not([data-cssvars="out"])'));if(o.__benchmark=R(),o.exclude=[C?'[data-cssvars]:not([data-cssvars=""])':'[data-cssvars="out"]',"link[disabled]:not([data-cssvars])",o.exclude].filter((function(e){return e})).join(","),o.variables=L(o.variables),c.forEach((function(e){var t="style"===e.nodeName.toLowerCase()&&e.__cssVars.text,r=t&&e.textContent!==e.__cssVars.text;t&&r&&(e.sheet&&(e.sheet.disabled=!1),e.setAttribute("data-cssvars",""))})),!C){var d=[].slice.call(o.rootElement.querySelectorAll('[data-cssvars="out"]'));d.forEach((function(e){var t=e.getAttribute("data-cssvars-group");t&&o.rootElement.querySelector('[data-cssvars="src"][data-cssvars-group="'.concat(t,'"]'))||e.parentNode.removeChild(e)})),x&&c.length<x&&(x=c.length,w.dom={})}}if("loading"!==document.readyState)if(h&&o.onlyLegacy){var m=!1;if(o.updateDOM){var v=o.rootElement.host||(o.rootElement===document?document.documentElement:o.rootElement);Object.keys(o.variables).forEach((function(e){var t=o.variables[e];m=m||t!==getComputedStyle(v).getPropertyValue(e),v.style.setProperty(e,t)}))}i(m)}else!O&&(o.shadowDOM||o.rootElement.shadowRoot||o.rootElement.host)?n({rootElement:b.rootElement,include:b.include,exclude:o.exclude,skipDisabled:!1,onSuccess:function(e,t,r){return!((t.sheet||{}).disabled&&!t.__cssVars)&&((e=((e=e.replace(S.cssComments,"").replace(S.cssMediaQueries,"")).match(S.cssVarDeclRules)||[]).join(""))||!1)},onComplete:function(e,t,r){l(e,{store:w.dom,onWarning:a}),O=!0,j(o)}}):(E=o.rootElement,n({rootElement:o.rootElement,include:o.include,exclude:o.exclude,skipDisabled:!1,onBeforeSend:o.onBeforeSend,onError:function(e,t,r){var n=e.responseURL||V(r,location.href),o=e.statusText?"(".concat(e.statusText,")"):"Unspecified Error"+(0===e.status?" (possibly CORS related)":"");s("CSS XHR Error: ".concat(n," ").concat(e.status," ").concat(o),t,e,n)},onSuccess:function(e,t,r){if((t.sheet||{}).disabled&&!t.__cssVars)return!1;var n="link"===t.nodeName.toLowerCase(),s="style"===t.nodeName.toLowerCase()&&e!==t.textContent,a=o.onSuccess(e,t,r);return e=void 0!==a&&!1===Boolean(a)?"":a||e,o.updateURLs&&(n||s)&&(e=T(e,r)),e},onComplete:function(t,r){var n=arguments.length>2&&void 0!==arguments[2]?arguments[2]:[],c=e({},w.dom,w.user);if(w.job={},n.forEach((function(e,t){var n=r[t];if(e.__cssVars=e.__cssVars||{},e.__cssVars.text=n,S.cssVars.test(n))try{var i=u(n,{preserveStatic:o.preserveStatic,removeComments:!0});l(i,{parseHost:Boolean(o.rootElement.host),store:w.dom,onWarning:a}),e.__cssVars.tree=i}catch(t){s(t.message,e)}})),e(w.job,w.dom),o.updateDOM?(e(w.user,o.variables),e(w.job,w.user)):(e(w.job,w.user,o.variables),e(c,o.variables)),y.job>0&&Boolean(Object.keys(w.job).length>Object.keys(c).length||Boolean(Object.keys(c).length&&Object.keys(w.job).some((function(e){return w.job[e]!==c[e]})))))D(o.rootElement),j(o);else{var d=[],m=[],v=!1;if(o.updateDOM&&y.job++,n.forEach((function(t,n){var i=!t.__cssVars.tree;if(t.__cssVars.tree)try{p(t.__cssVars.tree,e({},o,{variables:w.job,onWarning:a}));var c=f(t.__cssVars.tree);if(o.updateDOM){var u=r[n],l=S.cssVarFunc.test(u);if(t.getAttribute("data-cssvars")||t.setAttribute("data-cssvars","src"),c.length&&l){var g=t.getAttribute("data-cssvars-group")||++y.group,h=c.replace(/\s/g,""),b=o.rootElement.querySelector('[data-cssvars="out"][data-cssvars-group="'.concat(g,'"]'))||document.createElement("style");v=v||S.cssKeyframes.test(c),o.preserveStatic&&t.sheet&&(t.sheet.disabled=!0),b.hasAttribute("data-cssvars")||b.setAttribute("data-cssvars","out"),h===t.textContent.replace(/\s/g,"")?(i=!0,b&&b.parentNode&&(t.removeAttribute("data-cssvars-group"),b.parentNode.removeChild(b))):h!==b.textContent.replace(/\s/g,"")&&([t,b].forEach((function(e){e.setAttribute("data-cssvars-job",y.job),e.setAttribute("data-cssvars-group",g)})),b.textContent=c,d.push(c),m.push(b),b.parentNode||t.parentNode.insertBefore(b,t.nextSibling))}}else t.textContent.replace(/\s/g,"")!==c&&d.push(c)}catch(e){s(e.message,t)}i&&t.setAttribute("data-cssvars","skip"),t.hasAttribute("data-cssvars-job")||t.setAttribute("data-cssvars-job",y.job)})),x=o.rootElement.querySelectorAll('[data-cssvars]:not([data-cssvars="out"])').length,o.shadowDOM)for(var g,h=[].concat(o.rootElement).concat([].slice.call(o.rootElement.querySelectorAll("*"))),b=0;g=h[b];++b)if(g.shadowRoot&&g.shadowRoot.querySelector("style")){var C=e({},o,{rootElement:g.shadowRoot});j(C)}o.updateDOM&&v&&M(o.rootElement),E=!1,o.onComplete(d.join(""),m,JSON.parse(JSON.stringify(w.job)),R()-o.__benchmark),i(m.length)}}}));else document.addEventListener("DOMContentLoaded",(function e(r){j(t),document.removeEventListener("DOMContentLoaded",e)}))}}function A(e){function t(e){var t=r(e)&&e.hasAttribute("disabled"),n=(e.sheet||{}).disabled;return t||n}function r(e){return"link"===e.nodeName.toLowerCase()&&-1!==(e.getAttribute("rel")||"").indexOf("stylesheet")}function n(e){return"style"===e.nodeName.toLowerCase()}window.MutationObserver&&(C&&(C.disconnect(),C=null),(C=new MutationObserver((function(o){o.some((function(o){return function(n){var o=!1;if("attributes"===n.type&&r(n.target)&&!t(n.target)){var s="disabled"===n.attributeName,a="href"===n.attributeName,i="skip"===n.target.getAttribute("data-cssvars"),c="src"===n.target.getAttribute("data-cssvars");s?o=!i&&!c:a&&(i?n.target.setAttribute("data-cssvars",""):c&&D(e.rootElement,!0),o=!0)}return o}(o)||function(e){var t=!1;if("childList"===e.type){var r=n(e.target),o="out"===e.target.getAttribute("data-cssvars");t=r&&!o}return t}(o)||function(e){var o=!1;return"childList"===e.type&&(o=[].slice.call(e.addedNodes).some((function(e){var o=1===e.nodeType&&e.hasAttribute("data-cssvars"),s=n(e)&&S.cssVars.test(e.textContent);return!o&&(r(e)||s)&&!t(e)}))),o}(o)||function(t){var r=!1;return"childList"===t.type&&(r=[].slice.call(t.removedNodes).some((function(t){var r=1===t.nodeType,n=r&&"out"===t.getAttribute("data-cssvars"),o=r&&"src"===t.getAttribute("data-cssvars"),s=o;if(o||n){var a=t.getAttribute("data-cssvars-group"),i=e.rootElement.querySelector('[data-cssvars-group="'.concat(a,'"]'));o&&D(e.rootElement,!0),i&&i.parentNode.removeChild(i)}return s}))),r}(o)}))&&j(e)}))).observe(document.documentElement,{attributes:!0,attributeFilter:["disabled","href"],childList:!0,subtree:!0}))}function _(e){var t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:100;clearTimeout(k),k=setTimeout((function(){e.__benchmark=null,j(e)}),t)}function M(e){var t=["animation-name","-moz-animation-name","-webkit-animation-name"].filter((function(e){return getComputedStyle(document.body)[e]}))[0];if(t){for(var r=e.getElementsByTagName("*"),n=[],o="__CSSVARSPONYFILL-KEYFRAMES__",s=0,a=r.length;s<a;s++){var i=r[s];"none"!==getComputedStyle(i)[t]&&(i.style[t]+=o,n.push(i))}document.body.offsetHeight;for(var c=0,u=n.length;c<u;c++){var l=n[c].style;l[t]=l[t].replace(o,"")}}}function T(e,t){return(e.replace(S.cssComments,"").match(S.cssUrls)||[]).forEach((function(r){var n=r.replace(S.cssUrls,"$1"),o=V(n,t);e=e.replace(r,r.replace(n,o))})),e}function L(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{},t=/^-{2}/;return Object.keys(e).reduce((function(r,n){return r[t.test(n)?n:"--".concat(n.replace(/^-+/,""))]=e[n],r}),{})}function V(e){var t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:location.href,r=document.implementation.createHTMLDocument(""),n=r.createElement("base"),o=r.createElement("a");return r.head.appendChild(n),r.body.appendChild(o),n.href=t,o.href=e,o.href}function R(){return g&&(window.performance||{}).now?window.performance.now():(new Date).getTime()}function D(e){var t=arguments.length>1&&void 0!==arguments[1]&&arguments[1],r=[].slice.call(e.querySelectorAll('[data-cssvars="skip"],[data-cssvars="src"]'));r.forEach((function(e){return e.setAttribute("data-cssvars","")})),t&&(w.dom={})}j.reset=function(){for(var e in y.job=0,y.group=0,E=!1,C&&(C.disconnect(),C=null),x=0,k=null,O=!1,w)w[e]={}};var N=j,B=r(269),F=r.n(B);document.addEventListener("DOMContentLoaded",(()=>{N(),F()(null,{watchMQ:!0})}))}()}();
