@@ -15,7 +15,7 @@ class FilesystemTest extends WP_UnitTestCase {
 	 */
 	function test_get_file_contents() {
 		$svg  = '<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-activity"><polyline points="22 12 18 12 15 21 9 3 6 12 2 12"></polyline></svg>';
-		$file = \ystandard\Filesystem::file_get_contents(
+		$file = \ystandard\helper\Filesystem::file_get_contents(
 			get_template_directory() . '/library/feather/activity.svg'
 		);
 		$this->assertSame( $svg, $file );
